+++ com.sap.cloud.lm.sl.mta/src/main/java/com/sap/cloud/lm/sl/mta/model/v2/DeploymentDescriptor.java
@@ -10,18 +10,37 @@ import org.apache.commons.lang3.ObjectUtils;
 
 import com.sap.cloud.lm.sl.common.util.ListUtil;
 import com.sap.cloud.lm.sl.common.util.MapUtil;
+import com.sap.cloud.lm.sl.mta.model.Descriptor;
+import com.sap.cloud.lm.sl.mta.model.ElementContext;
 import com.sap.cloud.lm.sl.mta.model.ParametersContainer;
+import com.sap.cloud.lm.sl.mta.model.PropertiesContainer;
+import com.sap.cloud.lm.sl.mta.model.VisitableElement;
+import com.sap.cloud.lm.sl.mta.model.Visitor;
 import com.sap.cloud.lm.sl.mta.parsers.v2.DeploymentDescriptorParser;
 import com.sap.cloud.lm.sl.mta.util.YamlElement;
 import com.sap.cloud.lm.sl.mta.util.YamlElementOrder;
 
 @YamlElementOrder({ "id", "description", "version", "provider", "copyright", "schemaVersion", "parameters", "modules2", "resources2" })
-public class DeploymentDescriptor extends com.sap.cloud.lm.sl.mta.model.v1.DeploymentDescriptor implements ParametersContainer {
+public class DeploymentDescriptor implements Descriptor, VisitableElement, PropertiesContainer, ParametersContainer {
 
+    @YamlElement(DeploymentDescriptorParser.SCHEMA_VERSION)
+    private String schemaVersion;
+    @YamlElement(DeploymentDescriptorParser.ID)
+    private String id;
+    @YamlElement(DeploymentDescriptorParser.DESCRIPTION)
+    private String description;
+    @YamlElement(DeploymentDescriptorParser.VERSION)
+    private String version;
+    @YamlElement(DeploymentDescriptorParser.PROVIDER)
+    private String provider;
+    @YamlElement(DeploymentDescriptorParser.COPYRIGHT)
+    private String copyright;
     @YamlElement(DeploymentDescriptorParser.MODULES)
     private List<Module> modules2;
     @YamlElement(DeploymentDescriptorParser.RESOURCES)
     private List<Resource> resources2;
+    @YamlElement(DeploymentDescriptorParser.PROPERTIES)
+    private Map<String, Object> properties;
     @YamlElement(DeploymentDescriptorParser.PARAMETERS)
     private Map<String, Object> parameters;
 
@@ -29,11 +48,15 @@ public class DeploymentDescriptor extends com.sap.cloud.lm.sl.mta.model.v1.Deplo
 
     }
 
+    @Override
+    public String getId() {
+        return id;
+    }
+    
     public List<Module> getModules2() {
         return ListUtil.upcastUnmodifiable(getModules());
     }
 
-    @Override
     protected List<? extends Module> getModules() {
         return modules2;
     }
@@ -42,7 +65,6 @@ public class DeploymentDescriptor extends com.sap.cloud.lm.sl.mta.model.v1.Deplo
         return ListUtil.upcastUnmodifiable(getResources());
     }
 
-    @Override
     protected List<? extends Resource> getResources() {
         return resources2;
     }
@@ -61,8 +83,7 @@ public class DeploymentDescriptor extends com.sap.cloud.lm.sl.mta.model.v1.Deplo
         setModules(modules);
     }
 
-    @Override
-    protected void setModules(List<? extends com.sap.cloud.lm.sl.mta.model.v1.Module> modules) {
+    protected void setModules(List<? extends Module> modules) {
         this.modules2 = ListUtil.cast(modules);
     }
 
@@ -70,8 +91,7 @@ public class DeploymentDescriptor extends com.sap.cloud.lm.sl.mta.model.v1.Deplo
         setResources(resources);
     }
 
-    @Override
-    protected void setResources(List<? extends com.sap.cloud.lm.sl.mta.model.v1.Resource> resources) {
+    protected void setResources(List<? extends Resource> resources) {
         this.resources2 = ListUtil.cast(resources);
     }
 
@@ -84,8 +104,18 @@ public class DeploymentDescriptor extends com.sap.cloud.lm.sl.mta.model.v1.Deplo
     public void setParameters(Map<String, Object> parameters) {
         this.parameters = new LinkedHashMap<>(parameters);
     }
-
+    
     @Override
+    public void accept(ElementContext context, Visitor visitor) {
+        visitor.visit(context, this);
+        for (Module module : getModules()) {
+            module.accept(new ElementContext(module, context), visitor);
+        }
+        for (Resource resource : getResources()) {
+            resource.accept(new ElementContext(resource, context), visitor);
+        }
+    }
+
     public DeploymentDescriptor copyOf() {
         Builder result = new Builder();
         result.setId(getId());
@@ -108,13 +138,18 @@ public class DeploymentDescriptor extends com.sap.cloud.lm.sl.mta.model.v1.Deplo
         return result.build();
     }
 
-    public static class Builder extends com.sap.cloud.lm.sl.mta.model.v1.DeploymentDescriptor.Builder {
+    public static class Builder {
 
+        protected String schemaVersion;
+        protected String id;
+        protected String description;
+        protected String version;
+        protected String provider;
+        protected String copyright;
         protected List<Module> modules2;
         protected List<Resource> resources2;
         protected Map<String, Object> parameters;
 
-        @Override
         public DeploymentDescriptor build() {
             DeploymentDescriptor result = new DeploymentDescriptor();
             result.setId(id);
@@ -129,12 +164,35 @@ public class DeploymentDescriptor extends com.sap.cloud.lm.sl.mta.model.v1.Deplo
             return result;
         }
 
+        public void setSchemaVersion(String schemaVersion) {
+            this.schemaVersion = schemaVersion;
+        }
+
+        public void setId(String id) {
+            this.id = id;
+        }
+
+        public void setDescription(String description) {
+            this.description = description;
+        }
+
+        public void setVersion(String version) {
+            this.version = version;
+        }
+
+        public void setProvider(String provider) {
+            this.provider = provider;
+        }
+
+        public void setCopyright(String copyright) {
+            this.copyright = copyright;
+        }
+
         public void setModules2(List<Module> modules) {
             setModules(modules);
         }
 
-        @Override
-        protected void setModules(List<? extends com.sap.cloud.lm.sl.mta.model.v1.Module> modules) {
+        protected void setModules(List<? extends Module> modules) {
             this.modules2 = ListUtil.cast(modules);
         }
 
@@ -142,8 +200,7 @@ public class DeploymentDescriptor extends com.sap.cloud.lm.sl.mta.model.v1.Deplo
             setResources(resources);
         }
 
-        @Override
-        protected void setResources(List<? extends com.sap.cloud.lm.sl.mta.model.v1.Resource> resources) {
+        protected void setResources(List<? extends Resource> resources) {
             this.resources2 = ListUtil.cast(resources);
         }
 
@@ -151,11 +208,54 @@ public class DeploymentDescriptor extends com.sap.cloud.lm.sl.mta.model.v1.Deplo
             this.parameters = parameters;
         }
 
-        @Override
         public void setProperties(Map<String, Object> properties) {
             throw new UnsupportedOperationException();
         }
 
+    }
+
+    public String getProvider() {
+        return provider;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public String getVersion() {
+        return version;
+    }
+
+    public String getCopyright() {
+        return copyright;
+    }
+
+    public String getSchemaVersion() {
+        return schemaVersion;
+    }
+
+    public void setSchemaVersion(String schemaVersion) {
+        this.schemaVersion = schemaVersion;
+    }
+
+    public void setId(String id) {
+        this.id = id;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    public void setVersion(String version) {
+        this.version = version;
+    }
+
+    public void setProvider(String provider) {
+        this.provider = provider;
+    }
+
+    public void setCopyright(String copyright) {
+        this.copyright = copyright;
     }
 
 }
